+++ src/main.cpp
@@ -2,14 +2,14 @@
 
 // the main file, where the main loop is happening, and also where the file members are defined
 
-unsigned int file::inputLineCount = 0; // number of lines in input file
+unsigned long file::inputLineCount = 0; // number of lines in input file
 
 bool debug = false; // if compilation is being debugged (default: false)
 
-unsigned int file::asm_data; // integers for tracking sections in asm file
-unsigned int file::asm_bss;  
-unsigned int file::asm_text;
-unsigned int file::asm_func;
+unsigned long file::asm_data; // integers for tracking sections in asm file
+unsigned long file::asm_bss;
+unsigned long file::asm_text;
+unsigned long file::asm_func;
 
 std::string file::inputText; // input file
 std::vector<std::string> file::outputVector = { // prefix for asm file
@@ -177,7 +177,7 @@ int main(int argc, char **argv)
 	
 	long end = duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count(); // end timer
 	std::string ms = std::to_string((end - start) % 1000); // get miliseconds time difference
-	int len = ms.length(); // add zeros to have 4 char string for miliseconds
+    unsigned long len = ms.length(); // add zeros to have 4 char string for miliseconds
 	for(int i = 4; i > len; i--)
 		ms.insert(ms.begin(), '0'); // insert zeros
 	
@@ -236,7 +236,7 @@ void file::append_func(std::string line)
 }
 
 
-void file::add(std::string line, int position)
+void file::add(std::string line, unsigned long position)
 {
 	file::outputVector.insert(file::outputVector.begin() + position, line); // insert line of code into asm file
 }
